pyproject.toml
[project]
name = "hubspot-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chift>=0.1.76",
    "fastapi[standard]>=0.115.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
